requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
google-genai
python-dotenv
gunicorn
whitenoise
dj-database-url
PyPDF2
Pillow
